pyproject.toml
[project]
name = "re-imbursement-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
]
